feat: support browser types
